fix(worker): worker cpu id 类型 uint8_t -> uint32_t，多 NUMA 高编号 CPU(如 240-319)可绑；不绑核哨兵 128 -> UINT32_MAX

+++ src/hcom/transport/rdma/verbs/net_rdma_driver.cpp
@@ -357,7 +357,7 @@ NResult NetDriverRDMA::CreateWorkers()
     std::vector<uint16_t> workerGroups;
     std::vector<int16_t> flatWorkerCpus;
     std::vector<int16_t> workerThreadPriority;
-    std::vector<std::pair<uint8_t, uint8_t>> workerGroupCpus;
+    std::vector<std::pair<uint32_t, uint32_t>> workerGroupCpus;
 
     /* parse */
     if (!(NetFunc::NN_ParseWorkersGroups(mOptions.WorkGroups(), workerGroups)) ||
